+++ src/main/java/com/work/erpsystem/service/impl/ItemServiceImpl.java
@@ -1,13 +1,17 @@
 package com.work.erpsystem.service.impl;
 
+import com.work.erpsystem.exception.NoDBRecord;
 import com.work.erpsystem.model.ItemModel;
 import com.work.erpsystem.repository.ItemRepository;
 import com.work.erpsystem.service.ItemService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
+@Slf4j
 @Service
 public class ItemServiceImpl implements ItemService {
 
@@ -24,13 +28,32 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public ItemModel findById(Long itemModelId) {
-        return itemRepository.findById(itemModelId).orElse(null);
+    public List<ItemModel> findAll() {
+        return itemRepository.findAll();
     }
 
     @Override
-    public ItemModel findByName(String itemModelName) {
-        return itemRepository.findByItemName(itemModelName);
+    public ItemModel findById(Long itemModelId) throws NoDBRecord {
+        ItemModel itemModel = itemRepository.findById(itemModelId).orElse(null);
+
+        if (Objects.isNull(itemModel)) {
+            String exceptionMessage = "No such record in data base with id: %d";
+            throw new NoDBRecord(String.format(exceptionMessage, itemModelId));
+        }
+
+        return itemModel;
+    }
+
+    @Override
+    public ItemModel findByName(String itemModelName) throws NoDBRecord {
+        ItemModel itemModel = itemRepository.findByItemName(itemModelName);
+
+        if (Objects.isNull(itemModel)) {
+            String exceptionMessage = "No such record in data base with name: %s";
+            throw new NoDBRecord(String.format(exceptionMessage, itemModelName));
+        }
+
+        return itemModel;
     }
 
     @Override
@@ -39,21 +62,20 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public List<ItemModel> findAll() {
-        return itemRepository.findAll();
-    }
-
-    @Override
-    public ItemModel deleteById(Long itemModelId) {
+    public void deleteById(Long itemModelId) throws NoDBRecord {
         ItemModel itemModel = this.findById(itemModelId);
+
+        if (Objects.isNull(itemModel)) {
+            String exceptionMessage = "No such record in data base with id: %d";
+            throw new NoDBRecord(String.format(exceptionMessage, itemModelId));
+        }
+
         itemRepository.deleteById(itemModelId);
 
-        return itemModel;
     }
 
     @Override
-    public ItemModel delete(ItemModel itemModel) {
+    public void delete(ItemModel itemModel) {
         itemRepository.delete(itemModel);
-        return itemModel;
     }
 }
